feat: move model_v2 into a seperate crate
