+++ src/meta/model_v2/migration/src/main.rs
@@ -2,5 +2,5 @@ use sea_orm_migration::prelude::*;
 
 #[async_std::main]
 async fn main() {
-    cli::run_cli(model_migration::Migrator).await;
+    cli::run_cli(risingwave_meta_model_migration::Migrator).await;
 }
