+++ src/meta/model_v2/src/connection.rs
@@ -12,29 +12,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use risingwave_pb::catalog::connection::PbInfo;
+use risingwave_pb::catalog::PbConnection;
 use sea_orm::entity::prelude::*;
+use sea_orm::ActiveValue;
 
-use crate::model_v2::DatabaseId;
+use crate::{ConnectionId, PrivateLinkService};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "database")]
+#[sea_orm(table_name = "connection")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub database_id: DatabaseId,
-    #[sea_orm(unique)]
+    pub connection_id: ConnectionId,
     pub name: String,
+    pub info: PrivateLinkService,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
         belongs_to = "super::object::Entity",
-        from = "Column::DatabaseId",
+        from = "Column::ConnectionId",
         to = "super::object::Column::Oid",
         on_update = "NoAction",
         on_delete = "Cascade"
     )]
     Object,
+    #[sea_orm(has_many = "super::sink::Entity")]
+    Sink,
+    #[sea_orm(has_many = "super::source::Entity")]
+    Source,
 }
 
 impl Related<super::object::Entity> for Entity {
@@ -43,4 +50,30 @@ impl Related<super::object::Entity> for Entity {
     }
 }
 
+impl Related<super::sink::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::Sink.def()
+    }
+}
+
+impl Related<super::source::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::Source.def()
+    }
+}
+
 impl ActiveModelBehavior for ActiveModel {}
+
+impl From<PbConnection> for ActiveModel {
+    fn from(conn: PbConnection) -> Self {
+        let Some(PbInfo::PrivateLinkService(private_link_srv)) = conn.info else {
+            unreachable!("private link not provided.")
+        };
+
+        Self {
+            connection_id: ActiveValue::Set(conn.id as _),
+            name: ActiveValue::Set(conn.name),
+            info: ActiveValue::Set(PrivateLinkService(private_link_srv)),
+        }
+    }
+}
