+++ src/meta/model_v2/src/index.rs
@@ -14,43 +14,47 @@
 
 use sea_orm::entity::prelude::*;
 
-use crate::model_v2::I32Array;
+use crate::{ExprNodeArray, I32Array, IndexId, JobStatus, TableId};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "fragment")]
+#[sea_orm(table_name = "index")]
 pub struct Model {
-    #[sea_orm(primary_key)]
-    pub fragment_id: i32,
-    pub table_id: i32,
-    pub fragment_type_mask: i32,
-    pub distribution_type: String,
-    pub stream_node: Json,
-    pub vnode_mapping: Option<Json>,
-    pub state_table_ids: Option<I32Array>,
-    pub upstream_fragment_id: Option<I32Array>,
-    pub dispatcher_type: Option<String>,
-    pub dist_key_indices: Option<I32Array>,
-    pub output_indices: Option<I32Array>,
+    #[sea_orm(primary_key, auto_increment = false)]
+    pub index_id: IndexId,
+    pub name: String,
+    pub index_table_id: TableId,
+    pub primary_table_id: TableId,
+    pub index_items: ExprNodeArray,
+    pub original_columns: I32Array,
+    pub job_status: JobStatus,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
-    #[sea_orm(has_many = "super::actor::Entity")]
-    Actor,
     #[sea_orm(
         belongs_to = "super::object::Entity",
-        from = "Column::TableId",
+        from = "Column::IndexId",
         to = "super::object::Column::Oid",
         on_update = "NoAction",
         on_delete = "Cascade"
     )]
     Object,
-}
-
-impl Related<super::actor::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Actor.def()
-    }
+    #[sea_orm(
+        belongs_to = "super::table::Entity",
+        from = "Column::IndexTableId",
+        to = "super::table::Column::TableId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    Table2,
+    #[sea_orm(
+        belongs_to = "super::table::Entity",
+        from = "Column::PrimaryTableId",
+        to = "super::table::Column::TableId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    Table1,
 }
 
 impl Related<super::object::Entity> for Entity {
