+++ src/meta/model_v2/src/lib.rs
@@ -27,7 +27,6 @@ pub mod compaction_status;
 pub mod compaction_task;
 pub mod connection;
 pub mod database;
-pub mod ext;
 pub mod fragment;
 pub mod function;
 pub mod hummock_pinned_snapshot;
@@ -42,7 +41,6 @@ pub mod sink;
 pub mod source;
 pub mod system_parameter;
 pub mod table;
-pub mod trx;
 pub mod user;
 pub mod user_privilege;
 pub mod view;
