+++ src/meta/service/Cargo.toml
@@ -23,6 +23,7 @@ regex = "1"
 risingwave_common = { workspace = true }
 risingwave_connector = { workspace = true }
 risingwave_meta = { workspace = true }
+risingwave_meta_model_v2 = { workspace = true }
 risingwave_pb = { workspace = true }
 sea-orm = { version = "0.12.0", features = [
     "sqlx-mysql",
