+++ src/meta/src/lib.rs
@@ -42,7 +42,6 @@ pub mod error;
 pub mod hummock;
 pub mod manager;
 pub mod model;
-pub mod model_v2;
 pub mod rpc;
 pub mod serving;
 pub mod storage;
