--- src/meta/src/model_v2/migration/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "model_migration"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "model_migration"
-path = "src/lib.rs"
-
-[dependencies]
-async-std = { version = "1", features = ["attributes", "tokio1"] }
-uuid = { version = "1", features = ["v4"] }
-
-[dependencies.sea-orm-migration]
-version = "0.12.0"
-features = ["sqlx-mysql", "sqlx-postgres", "sqlx-sqlite", "runtime-tokio-native-tls", "with-uuid"]
